+++ History.md
@@ -1,6 +1,18 @@
-## v0.9.3
+## v0.9.2.6 “InviteScope”
+
+* Added new invite features (thanks [@callmephilip](https://github.com/callmephilip)!)
+* Changed `navItems` to `primaryNav` and added `secondaryNav`. 
+* Added new `themeSettings` object for storing theme-level settings.
+* Notifications is now a nav menu item. 
+* Renamed `comments` to `commentsCount` on `Post` model.
+* Now tracking list of commenters `_id`s on `Post` model.
+* Rerun interrupted migrations. 
+
+## v0.9.2.5 “AccountScope”
 
 * Now keeping track of upvoted/downvoted posts & comments.
+* Using [accounts-entry](https://github.com/Differential/accounts-entry/) for accounts stuff.
+* Split out part of the Hubble theme into a new “base” theme.
 
 ## v0.9.2 “MailScope”